This is a (rather) large, messy commit. Goals:
* It moves the logic for Memory and RegStorage out of the context. A
context is now generic over 2 types: memory(which implements trait Memory)
and register storage(which implements RegStore).

* Memory: Since we define a trait for memory to be implemented, this
allows us to add different models for symbolic memory. In seg_mem.rs, I
have started with the basic idea for implementing the segment-tree
backed idea is going to be worked upon soon. In qword_mem.rs, the
initial memory model of chunks of size of a qword(8 bytes) is present.

* RegStore: This allows us to decouple register store/load logic from
the context core.

* The directory structure is kept such that we have ->
    - module/
        - trait_definition
        - [trait implementation - 1]
        - [trait implementation - 2]
        ..
    - utils/

+++ src/engine/rune.rs
@@ -141,7 +141,7 @@ where Ctx: Context<IFn=qf_abv::QF_ABV_Fn>,
                 return Ok(None);
             }
             Token::EPoke(size) => {
-                self.ctx.mem_write(l_op.unwrap(), r_op.unwrap(), size as u64);
+                self.ctx.mem_write(l_op.unwrap(), r_op.unwrap(), size as usize);
                 return Ok(None);
             }
             Token::ENop => return Ok(None),
@@ -150,7 +150,7 @@ where Ctx: Context<IFn=qf_abv::QF_ABV_Fn>,
 
         let result = match token {
             Token::EPeek(size) => {
-                self.ctx.mem_read(l_op.unwrap(), size as u64)
+                self.ctx.mem_read(l_op.unwrap(), size as usize)
             }
             Token::ECmp | Token::ELt | Token::EGt => {
                 // This case is a bit different as we want the result to be a bitvector rather
